AWS/RDS/telehealth-app/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1